git subrepo commit (merge) uno
subrepo:
  subdir:   "uno"
  merged:   "888f12be"
upstream:
  origin:   "https://github.com/prrvchr/uno.git"
  branch:   "main"
  commit:   "ee38b1eb"
git-subrepo:
  version:  "0.4.3"
  origin:   "https://github.com/ingydotnet/git-subrepo.git"
  commit:   "2f68596"

+++ uno/lib/python/setuptools/command/bdist_wheel.py
@@ -9,31 +9,31 @@ from __future__ import annotations
 import os
 import re
 import shutil
+import stat
 import struct
 import sys
 import sysconfig
 import warnings
-from collections.abc import Iterable, Sequence
-from email.generator import BytesGenerator
+from email.generator import BytesGenerator, Generator
+from email.policy import EmailPolicy
 from glob import iglob
-from typing import Literal, cast
+from shutil import rmtree
+from typing import TYPE_CHECKING, Callable, Iterable, Literal, Sequence, cast
 from zipfile import ZIP_DEFLATED, ZIP_STORED
 
 from packaging import tags, version as _packaging_version
+from wheel.metadata import pkginfo_to_metadata
 from wheel.wheelfile import WheelFile
 
-from .. import Command, __version__, _shutil
+from .. import Command, __version__
+from .._normalization import safer_name
 from ..warnings import SetuptoolsDeprecationWarning
 from .egg_info import egg_info as egg_info_cls
 
 from distutils import log
 
-
-def safe_name(name: str) -> str:
-    """Convert an arbitrary string to a standard distribution name
-    Any runs of non-alphanumeric/. characters are replaced with a single '-'.
-    """
-    return re.sub("[^A-Za-z0-9.]+", "-", name)
+if TYPE_CHECKING:
+    from _typeshed import ExcInfo
 
 
 def safe_version(version: str) -> str:
@@ -58,7 +58,7 @@ def _is_32bit_interpreter() -> bool:
 
 
 def python_tag() -> str:
-    return f"py{sys.version_info.major}"
+    return f"py{sys.version_info[0]}"
 
 
 def get_platform(archive_root: str | None) -> str:
@@ -104,11 +104,19 @@ def get_abi_tag() -> str | None:
     impl = tags.interpreter_name()
     if not soabi and impl in ("cp", "pp") and hasattr(sys, "maxunicode"):
         d = ""
+        m = ""
         u = ""
         if get_flag("Py_DEBUG", hasattr(sys, "gettotalrefcount"), warn=(impl == "cp")):
             d = "d"
 
-        abi = f"{impl}{tags.interpreter_version()}{d}{u}"
+        if get_flag(
+            "WITH_PYMALLOC",
+            impl == "cp",
+            warn=(impl == "cp" and sys.version_info < (3, 8)),
+        ) and sys.version_info < (3, 8):
+            m = "m"
+
+        abi = f"{impl}{tags.interpreter_version()}{d}{m}{u}"
     elif soabi and impl == "cp" and soabi.startswith("cpython"):
         # non-Windows
         abi = "cp" + soabi.split("-")[1]
@@ -133,12 +141,23 @@ def get_abi_tag() -> str | None:
     return abi
 
 
-def safer_name(name: str) -> str:
-    return safe_name(name).replace("-", "_")
-
-
 def safer_version(version: str) -> str:
     return safe_version(version).replace("-", "_")
+
+
+def remove_readonly(
+    func: Callable[..., object],
+    path: str,
+    excinfo: ExcInfo,
+) -> None:
+    remove_readonly_exc(func, path, excinfo[1])
+
+
+def remove_readonly_exc(
+    func: Callable[..., object], path: str, exc: BaseException
+) -> None:
+    os.chmod(path, stat.S_IWRITE)
+    func(path)
 
 
 class bdist_wheel(Command):
@@ -184,7 +203,9 @@ class bdist_wheel(Command):
         (
             "compression=",
             None,
-            f"zipfile compression (one of: {', '.join(supported_compressions)}) [default: 'deflated']",
+            "zipfile compression (one of: {}) [default: 'deflated']".format(
+                ", ".join(supported_compressions)
+            ),
         ),
         (
             "python-tag=",
@@ -216,7 +237,7 @@ class bdist_wheel(Command):
 
     def initialize_options(self) -> None:
         self.bdist_dir: str | None = None
-        self.data_dir = ""
+        self.data_dir: str | None = None
         self.plat_name: str | None = None
         self.plat_tag: str | None = None
         self.format = "zip"
@@ -434,7 +455,7 @@ class bdist_wheel(Command):
             shutil.copytree(self.dist_info_dir, distinfo_dir)
             # Egg info is still generated, so remove it now to avoid it getting
             # copied into the wheel.
-            _shutil.rmtree(self.egginfo_dir)
+            shutil.rmtree(self.egginfo_dir)
         else:
             # Convert the generated egg-info into dist-info.
             self.egg2dist(self.egginfo_dir, distinfo_dir)
@@ -452,14 +473,17 @@ class bdist_wheel(Command):
         # Add to 'Distribution.dist_files' so that the "upload" command works
         getattr(self.distribution, "dist_files", []).append((
             "bdist_wheel",
-            f"{sys.version_info.major}.{sys.version_info.minor}",
+            "{}.{}".format(*sys.version_info[:2]),  # like 3.7
             wheel_path,
         ))
 
         if not self.keep_temp:
             log.info(f"removing {self.bdist_dir}")
             if not self.dry_run:
-                _shutil.rmtree(self.bdist_dir)
+                if sys.version_info < (3, 12):
+                    rmtree(self.bdist_dir, onerror=remove_readonly)
+                else:
+                    rmtree(self.bdist_dir, onexc=remove_readonly_exc)
 
     def write_wheelfile(
         self, wheelfile_base: str, generator: str = f"setuptools ({__version__})"
@@ -498,7 +522,7 @@ class bdist_wheel(Command):
             # Setuptools has resolved any patterns to actual file names
             return self.distribution.metadata.license_files or ()
 
-        files = set[str]()
+        files: set[str] = set()
         metadata = self.distribution.get_option_dict("metadata")
         if setuptools_major_version >= 42:
             # Setuptools recognizes the license_files option but does not do globbing
@@ -543,7 +567,7 @@ class bdist_wheel(Command):
         def adios(p: str) -> None:
             """Appropriately delete directory, file or link."""
             if os.path.exists(p) and not os.path.islink(p) and os.path.isdir(p):
-                _shutil.rmtree(p)
+                shutil.rmtree(p)
             elif os.path.exists(p):
                 os.unlink(p)
 
@@ -565,30 +589,42 @@ class bdist_wheel(Command):
 
             raise ValueError(err)
 
-        # .egg-info is a directory
-        pkginfo_path = os.path.join(egginfo_path, "PKG-INFO")
+        if os.path.isfile(egginfo_path):
+            # .egg-info is a single file
+            pkg_info = pkginfo_to_metadata(egginfo_path, egginfo_path)
+            os.mkdir(distinfo_path)
+        else:
+            # .egg-info is a directory
+            pkginfo_path = os.path.join(egginfo_path, "PKG-INFO")
+            pkg_info = pkginfo_to_metadata(egginfo_path, pkginfo_path)
 
-        # ignore common egg metadata that is useless to wheel
-        shutil.copytree(
-            egginfo_path,
-            distinfo_path,
-            ignore=lambda x, y: {
-                "PKG-INFO",
-                "requires.txt",
-                "SOURCES.txt",
-                "not-zip-safe",
-            },
+            # ignore common egg metadata that is useless to wheel
+            shutil.copytree(
+                egginfo_path,
+                distinfo_path,
+                ignore=lambda x, y: {
+                    "PKG-INFO",
+                    "requires.txt",
+                    "SOURCES.txt",
+                    "not-zip-safe",
+                },
+            )
+
+            # delete dependency_links if it is only whitespace
+            dependency_links_path = os.path.join(distinfo_path, "dependency_links.txt")
+            with open(dependency_links_path, encoding="utf-8") as dependency_links_file:
+                dependency_links = dependency_links_file.read().strip()
+            if not dependency_links:
+                adios(dependency_links_path)
+
+        pkg_info_path = os.path.join(distinfo_path, "METADATA")
+        serialization_policy = EmailPolicy(
+            utf8=True,
+            mangle_from_=False,
+            max_line_length=0,
         )
-
-        # delete dependency_links if it is only whitespace
-        dependency_links_path = os.path.join(distinfo_path, "dependency_links.txt")
-        with open(dependency_links_path, encoding="utf-8") as dependency_links_file:
-            dependency_links = dependency_links_file.read().strip()
-        if not dependency_links:
-            adios(dependency_links_path)
-
-        metadata_path = os.path.join(distinfo_path, "METADATA")
-        shutil.copy(pkginfo_path, metadata_path)
+        with open(pkg_info_path, "w", encoding="utf-8") as out:
+            Generator(out, policy=serialization_policy).flatten(pkg_info)
 
         for license_path in self.license_paths:
             filename = os.path.basename(license_path)
